Files moved to smoke folder
